Draft: Storage Module

See merge request heliogrim-workgroup/engine!4

+++ src/Editor.Render/Renderer/EditorRenderer.cpp
@@ -1,5 +1,6 @@
 #include "EditorRenderer.hpp"
 
+#include <utility>
 #include <Engine.Common/Make.hpp>
 #include <Engine.Common/Meta/TypeId.hpp>
 #include <Engine.GFX.Render.Predefined/Symbols/SceneColor.hpp>
@@ -7,8 +8,8 @@
 #include <Engine.GFX.Render.Subpass/Impl/TmpEndPass.hpp>
 #include <Engine.GFX.RenderGraph/CompileGraph.hpp>
 #include <Engine.GFX.RenderGraph/Builder/Builder.hpp>
+#include <Engine.GFX.RenderGraph/Node/Compile/CompileSubpassNode.hpp>
 #include <Engine.GFX.RenderGraph/Node/Runtime/BarrierNode.hpp>
-#include <Engine.GFX.RenderGraph/Node/Compile/CompileSubPassNode.hpp>
 #include <Engine.GFX.RenderGraph/Pass/CompilePassContext.hpp>
 #include <Engine.Reflow.Render/ReflowPass.hpp>
 
@@ -21,7 +22,7 @@ EditorRenderer::EditorRenderer(
 	mref<nmpt<engine::gfx::memory::GlobalPooledAllocator>> globalGfxAllocator_
 ) noexcept :
 	Renderer(
-		Guid { "__Static__Proxy"_typeId.data, 0u, 0u, "Editor-Renderer"_typeId.data },
+		Guid { static_cast<u32>("__Static__Proxy"_typeId.data), 0u, 0u, "Editor-Renderer"_typeId.data },
 		"Editor-Renderer",
 		makeCompileGraph(),
 		nullptr,
@@ -56,7 +57,7 @@ uptr<engine::render::graph::CompileGraph> EditorRenderer::makeCompileGraph() noe
 	beginPass->setSubPassBuilder(
 		[](cref<graph::CompilePassContext> ctx_) -> uptr<graph::SubPassNodeBase> {
 			auto node = ctx_.getGraphNodeAllocator()
-			                ->allocate<graph::SubPassNode<DummyProvider>>();
+				->allocate<graph::SubPassNode<DummyProvider>>();
 			return node;
 		}
 	);
@@ -64,7 +65,7 @@ uptr<engine::render::graph::CompileGraph> EditorRenderer::makeCompileGraph() noe
 	uiPass->setSubPassBuilder(
 		[](cref<graph::CompilePassContext> ctx_) -> uptr<graph::SubPassNodeBase> {
 			auto node = ctx_.getGraphNodeAllocator()
-			                ->allocate<graph::SubPassNode<engine::reflow::render::ReflowPass>>();
+				->allocate<graph::SubPassNode<engine::reflow::render::ReflowPass>>();
 			return node;
 		}
 	);
@@ -72,7 +73,7 @@ uptr<engine::render::graph::CompileGraph> EditorRenderer::makeCompileGraph() noe
 	endPass->setSubPassBuilder(
 		[](cref<graph::CompilePassContext> ctx_) -> uptr<graph::SubPassNodeBase> {
 			auto node = ctx_.getGraphNodeAllocator()
-			                ->allocate<graph::SubPassNode<TmpEndPass>>(vk::ImageLayout::ePresentSrcKHR);
+				->allocate<graph::SubPassNode<TmpEndPass>>(vk::ImageLayout::ePresentSrcKHR);
 			return node;
 		}
 	);
