+++ src/Engine.GFX.Render.Predefined/Effects/Skybox.cpp
@@ -156,7 +156,7 @@ static smr<AccelerationEffect> build_test_effect() {
 
 	/**/
 
-	Guid guid { "__Test__Proxy"_typeId.data, 0, 0, "SkyboxEffect"_typeId.data };
+	Guid guid { static_cast<u32>("__Test__Proxy"_typeId.data), 0, 0, "SkyboxEffect"_typeId.data };
 
 	return make_smr<AccelerationEffect>(
 		std::move(guid),
